# [#148409] Reduce number of DB queries on Facility timeline page