--- a/app/controllers/facility_reservations_controller.rb
+++ b/app/controllers/facility_reservations_controller.rb
@@ -158,9 +158,10 @@ class FacilityReservationsController < ApplicationController
 
   def timeline
     @display_datetime = parse_usa_date(params[:date]) || Time.current.beginning_of_day
-    @schedules = current_facility.schedules.active
-                 .includes(facility_visible_products: [:alert, :relay])
-                 .order(:name)
+    @schedules = current_facility.schedules
+                                 .active
+                                 .includes(facility_visible_products: [:alert, :relay])
+                                 .order(:name)
   end
 
   protected