--- a/app/models/schedule.rb
+++ b/app/models/schedule.rb
@@ -2,30 +2,17 @@
 
 class Schedule < ApplicationRecord
 
-  # Associations
-  # --------
   belongs_to :facility
 
   has_many :products, class_name: "Instrument"
   has_many :reservations, through: :products
   has_many :admin_reservations, through: :products
+  has_many :publicly_visible_products, -> { active }, class_name: "Instrument"
+  has_many :facility_visible_products, -> { not_archived }, class_name: "Instrument"
 
-  # Validations
-  # --------
   validates_presence_of :facility
 
   scope :active, -> { where(id: Product.not_archived.with_schedule.select(:schedule_id)).order(:name) }
-
-  # Instance methods
-  # --------
-
-  def publicly_visible_products
-    products.active
-  end
-
-  def facility_visible_products
-    products.not_archived
-  end
 
   def shared?
     products.count > 1