--- a/app/views/facility_reservations/_schedule.html.haml
+++ b/app/views/facility_reservations/_schedule.html.haml
@@ -1,4 +1,4 @@
-- products = public_timeline? ? schedule.publicly_visible_products : schedule.facility_visible_products
+- products = schedule.facility_visible_products
 - if products.count > 1
   .timeline-schedule
     %h3= schedule.name