--- a/app/views/facility_reservations/timeline.html.haml
+++ b/app/views/facility_reservations/timeline.html.haml
@@ -29,4 +29,4 @@
 
 .clear
 .span12.timeline-wrapper
-  = render partial: "shared/timeline/schedule", collection: @schedules, as: :schedule
+  = render partial: "schedule", collection: @schedules, as: :schedule