--- a/app/views/reservations/_schedule.html.haml
+++ b/app/views/reservations/_schedule.html.haml
@@ -1,0 +1,8 @@
+- products = schedule.publicly_visible_products
+- if products.count > 1
+  .timeline-schedule
+    %h3= schedule.name
+    = render partial: "shared/timeline/instrument", collection: products, as: :instrument
+- else
+  .timeline-single
+    = render partial: "shared/timeline/instrument", collection: products, as: :instrument