--- a/app/views/reservations/_schedule.html.haml
+++ b/app/views/reservations/_schedule.html.haml
@@ -1,8 +1,7 @@
-- products = schedule.publicly_visible_products
-- if products.count > 1
+- if schedule.publicly_visible_products.size > 1
   .timeline-schedule
     %h3= schedule.name
-    = render partial: "shared/timeline/instrument", collection: products, as: :instrument
+    = render partial: "shared/timeline/instrument", collection: schedule.publicly_visible_products, as: :instrument
 - else
   .timeline-single
-    = render partial: "shared/timeline/instrument", collection: products, as: :instrument
+    = render partial: "shared/timeline/instrument", collection: schedule.publicly_visible_products, as: :instrument