--- a/app/views/reservations/new.html.haml
+++ b/app/views/reservations/new.html.haml
@@ -21,7 +21,7 @@
 
 - if @instrument.offline?
   %p.alert.alert-danger= text("instruments.offline.notice")
-- elsif @instrument.has_alert?
+- elsif @instrument.alert
   %p.alert.alert-warning= text("instruments.schedule.instrument_alert_is_active", note: @instrument.alert.note)
 
 = render "products_common/description", product: @instrument