--- a/app/views/reservations/public_timeline.html.haml
+++ b/app/views/reservations/public_timeline.html.haml
@@ -18,4 +18,4 @@
 
 .row
   .span12.timeline-wrapper
-    = render partial: "shared/timeline/schedule", collection: @schedules, as: :schedule
+    = render partial: "schedule", collection: @schedules, as: :schedule