--- a/app/views/shared/timeline/_relay_switch.html.haml
+++ b/app/views/shared/timeline/_relay_switch.html.haml
@@ -1,3 +1,0 @@
-.relay_checkbox
-  = check_box_tag "relay[#{instrument.id}]", 1, false, :disabled => true, :"data-relay-url" => facility_instrument_switch_path(instrument.facility, instrument)
-  .loading_box